Change counts type uint32_t to uint16_t

+++ iv/aero/ast.h
@@ -148,7 +148,7 @@ class CharacterAtom : public Atom {
 class RangeAtom : public Atom {
  public:
   explicit RangeAtom(bool inverted, bool singles,
-                     uint32_t counts, Ranges* ranges)
+                     uint16_t counts, Ranges* ranges)
     : inverted_(inverted)
     , singles_(singles)
     , counts_(counts)
@@ -165,12 +165,12 @@ class RangeAtom : public Atom {
   }
   bool inverted() const { return inverted_; }
   bool singles() const { return singles_; }
-  uint32_t counts() const { return counts_; }
+  uint16_t counts() const { return counts_; }
   DECLARE_DERIVED_NODE_TYPE(RangeAtom)
  private:
   bool inverted_;
   bool singles_;
-  uint32_t counts_;
+  uint16_t counts_;
   Ranges* ranges_;
 };
